requirements.txt
flask
flask-jsonpify
flask-pymongo
celery[redis]
redis
influxdb
gevent-socketio
